Adding tests for azure subscription, apiversion, and url encoding

+++ AutoRest/Generators/CSharp/Azure.CSharp.Tests/RegenerateExpectedClasses.ps1
@@ -5,6 +5,7 @@ $tests =
     "SwaggerBat\Report.Cs"="..\..\AcceptanceTests\swagger\azure-report.json";
     "SwaggerBat\ResourceFlattening.Cs"="..\..\AcceptanceTests\swagger\resource-flattening.json";
     "SwaggerBat\Head.Cs"="..\..\AcceptanceTests\swagger\head.json";
+    "SwaggerBat\AzureSpecials.Cs"="..\..\AcceptanceTests\swagger\azure-special-properties.json";
 }
 
 # TODO: direct reference to AutoRest .Net objects and methods leaves handles to the binaries open. Invoke the command line instead.
@@ -19,7 +20,10 @@ foreach ($test in $tests.GetEnumerator())
     $settings.OutputDirectory = "$PSScriptRoot\Expected\$($test.Key)"
     $settings.Input = "$PSScriptRoot\$($test.Value)"
     $settings.Header = "NONE"
-    Remove-Item "$($settings.OutputDirectory)" -Recurse -Force
+    if (Test-Path "$($settings.OutputDirectory)") 
+    {
+        Remove-Item "$($settings.OutputDirectory)" -Recurse -Force
+    }
     Write-Output "Generating $($test.Value)"
     [Microsoft.Rest.Generator.AutoRest]::Generate($settings)
     #&"$PSScriptRoot\..\..\..\..\binaries\net45\AutoRest.exe" -Modeler Swagger -CodeGenerator $flavor -OutputDirectory "$PSScriptRoot\Expected" -Namespace "$namespace" -Input "$PSScriptRoot\$($test.Value)" -Header NONE
